+++ CHANGELOG.md
@@ -5,7 +5,12 @@
 ### Implemented enhancements
 
 - GP2-1190 - Step 1 of 1 in removing wagtail-personalisation: unplug it from live code
+- GP2-1484 - Add override option for root link in breadcrumbs for ex-V1 pages
+- GP2-1177 - Getting paid - Payment methods element
+- GP2-1469 - Add table block to GuidancePage.body
+- GP2-1468 - Add InfoPage to CMS, to handle ex-BAU legal/guidance/terms content
 - GP2-1240 - Mark sections as complete
+- GP2-1344 - added trading bloc logic to search case study
 - GP2-1425 - Cost and pricing - Estimate element
 - GP2-1435 - Split data retrieval on 'Where to export' data table
 - GP2-1426 - Cost and pricing - Change to the title of the tool-tip
@@ -15,6 +20,7 @@
 - GP2-1411 - Refactor ArticlePage to use a StreamField with new PullQuoteBlock
 
 ### Fixed bugs
+- GP2-1361 - Compare markets - align cookies to user.
 - GP2-1397 - Render HTML definitions in product classifier.
 
 
@@ -105,6 +111,7 @@
 - No ticket - Added missing population url
 - GP2-1310 - Data not loading for if made more than one country is compared
 - GP2-1067 - Fixed LessonCompleted component
+- NOTICKET - Fix getting paid bug
 
 ## [1.0.1](https://github.com/uktrade/great-cms/releases/tag/1.0.1)